# feat: add support for case insensitive env names