--- a/docs/examples/settings_case_insensitive.py
+++ b/docs/examples/settings_case_insensitive.py
@@ -1,0 +1,8 @@
+from pydantic import BaseSettings
+
+
+class Settings(BaseSettings):
+    redis_host = 'localhost'
+
+    class Config:
+        case_insensitive = False