--- a/pydantic/env_settings.py
+++ b/pydantic/env_settings.py
@@ -21,7 +21,9 @@ class BaseSettings(BaseModel):
     """
     Base class for settings, allowing values to be overridden by environment variables.
 
-    Environment variables must be upper case. Eg. to override foobar, `export APP_FOOBAR="whatever"`.
+    Environment variables must be upper case and prefixed by APP_ by default. Eg. to override foobar,
+    `export APP_FOOBAR="whatever"`. To change this behaviour set Config options case_insensitive and
+    env_prefix.
 
     This is useful in production for secrets you do not wish to save in code, it places nicely with docker(-compose),
     Heroku and any 12 factor app design.
@@ -36,12 +38,23 @@ class BaseSettings(BaseModel):
         Substitute environment variables into values.
         """
         d = {}
+
+        if self.__config__.case_insensitive:
+            env_vars = {k.lower(): v for (k, v) in os.environ.items()}
+        else:
+            env_vars = dict(os.environ)
+
         for field in self.__fields__.values():
             if field.has_alias:
                 env_name = field.alias
             else:
                 env_name = self.__config__.env_prefix + field.name.upper()
-            env_var = os.getenv(env_name, None)
+
+            if self.__config__.case_insensitive:
+                env_var = env_vars.get(env_name.lower(), None)
+            else:
+                env_var = env_vars.get(env_name, None)
+
             if env_var:
                 if _complex_field(field):
                     try:
@@ -56,3 +69,4 @@ class BaseSettings(BaseModel):
         validate_all = True
         ignore_extra = False
         arbitrary_types_allowed = True
+        case_insensitive = False