--- a/pyproject.toml
+++ b/pyproject.toml
@@ -1,4 +1,0 @@
-[tool.black]
-skip-string-normalization = true
-line-length = 120
-py36 = true