feat: add support for case insensitive env names


 ## Change Summary

 <!-- Please give a short summary of the changes. -->

 `BaseSettings` can now be configured to read from os.environ in a
 case-insensitive way.

 ## Related issue number

 <!-- Are there any issues opened that will be resolved by merging this change? -->
Arguably will close #277. Doesn't quite meet the exact idea as proposed there, however it achieves the end goal that motivated #277, I think.

 ## Performance Changes

 pydantic cares about performance, if there's any risk performance changed on this PR,
 please run `make benchmark-pydantic` before and after the change:
 * before: **pydantic best=32.086μs/iter avg=33.792μs/iter stdev=1.696μs/iter**
 * after: **pydantic best=31.856μs/iter avg=32.579μs/iter stdev=0.731μs/iter**

 ## Checklist
 
* [x] Unit tests for the changes exist
* [x] Tests pass on CI and coverage remains at 100%
* [x] Documentation reflects the changes
* [x] No performance deterioration (if applicable)
* [x] `HISTORY.rst` has been updated
  * if this is the first change since a release, please add a new section
  * include the issue number or this pull request number `#<number>`
  * if you're not a regular contributer please include your github username `@whatever`